third_party/WebKit/LayoutTests/fast/canvas/linearGradient-infinite-values-expected.txt

This test checks createLinearGradient with infinite values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createLinearGradient(0, 0, 100, NaN) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(0, 0, 100, Infinity) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(0, 0, 100, -Infinity) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(0, 0, NaN, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(0, 0, Infinity, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(0, 0, -Infinity, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(0, NaN, 100, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(0, Infinity, 100, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(0, -Infinity, 100, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(NaN, 0, 100, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(Infinity, 0, 100, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.createLinearGradient(-Infinity, 0, 100, 100) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS successfullyParsed is true

TEST COMPLETE
